relocate things to Tetrium_Windowing and Tetrium_Config.cpp
